refactor(structure): restructure the folder
