+++ src/core/services/user/user-mock.service.ts
@@ -1,8 +1,8 @@
 import { BehaviorSubject, Observable } from "rxjs";
 
-import { HttpClient } from "@/libs/http/http-client";
+import { UserDomain } from "@/domain/user.domain";
 
-import { UserDomain } from "@/entities/user.domain";
+import { HttpClient } from "@/libs/http/http-client";
 
 import { UserService } from "./user.service";
