+++ src/core/usecases/header/header.usecase.ts
@@ -1,7 +1,7 @@
 import { Observable } from "rxjs";
 
-import { HeaderPresenter } from "@/presenters/header.presenter";
 import { UserService } from "@/services/user/user.service";
+import { HeaderPresenter } from "@/presenters/header.presenter";
 
 interface HeaderUsecaseI {
   userService: UserService;
