+++ tsconfig.json
@@ -16,15 +16,17 @@
     "incremental": true,
     "types": ["@testing-library/jest-dom"],
     "paths": {
-      "@/data/*": ["./src/core/data/*"],
-      "@/dataStore/*": ["./src/core/dataStore/*"],
-      "@/entities/*": ["./src/core/domain/entities/*"],
-      "@/presenters/*": ["./src/core/domain/presenters/*"],
-      "@/useCases/*": ["./src/core/domain/useCases/*"],
-      "@/hooks/*": ["./src/core/hooks/*"],
-      "@/libs/*": ["./src/core/libs/*"],
+      "@/domain/*": ["./src/core/entities/domain/*"],
+      "@/presenters/*": ["./src/core/entities/presenters/*"],
+      "@/entities/*": ["./src/core/entities/*"],
+      "@/usecases/*": ["./src/core/usecases/*"],
       "@/services/*": ["./src/core/services/*"],
-      "@/ui/*": ["./src/core/ui/*"],
+      "@/hooks/*": ["./src/hooks/*"],
+      "@/libs/*": ["./src/libs/*"],
+      "@/utils/*": ["./src/utils/*"],
+      "@/components/*": ["./src/components/*"],
+      "@/pages/*": ["./src/pages/*"],
+      "@/styles/*": ["./src/styles/*"],
       "@/*": ["./src/*"]
     }
   },
